+++ main_scripts/PARAMETERS/MakeParametersForPE.py
@@ -9,7 +9,7 @@ if __name__ == "__main__":
     keys = sorted(d.keys())
     
     with open(out_path, 'w') as file:
-        ### this part is only for correlation
+        ### FIXME: this part is only for correlation
         syn_path = '../CORRELATIONanalysis/synonims.tsv'
         names = []
         with open(syn_path, 'r') as syn:
@@ -21,7 +21,7 @@ if __name__ == "__main__":
                     names.append(GTRD_name)
         ###
         for key in keys:
-            ###
+            ### FIXME: this part is only for correlation
             name = key.split('!')[0]
             if name not in names:
                 print('Skipping, not in synonims..')
