+++ main_scripts/SNPcalling/__init__.py
@@ -1,0 +1,1 @@
+#from . import *
